requirements.txt
streamlit
google-generativeai
pypdf2
langchain
python-docx